Setting up pre-commit
This makes writing nicer code easier

set it up with `pre-commit install` (you need to install the pre-commit
package first) and then it automatically lints and formats before you
commit.

To start off I've run formatting on everything. Obviously it still fails
typing and linting checks, but we can either ignore that or fix it
incrementally, depending on preference.