Add behaviour recording and logging for Procgen and Atari
We use a wrapper around the environment to record the frames and wrap
them into videos, which can then be retrieved at the end of evaluation
and logged. Note the currently the logging might mess up the steps a
bit, I'm uncertain.

To satisfy linting/typing, also type some things and remove unused
imports.