add min-max logging
I think it makes more sense (especially when we're not building on
PPO) to normalise rewards based on minimum and maximum scores in games,
rather than PPO score. Note that the 'minimum' score isn't the score
of a totally random agent, but rather the score of an agent trained with
masked out observations, so it's the minimum score trivially achievable
in each environment. These numbers are taken from the OpenAI Procgen paper.